+++ src/routes/tools.js
@@ -95,10 +95,10 @@ export let tools = {
         link: '/robots-txt-generator',
         description: "The Robots.txt Generator is a must-have tool for developers to create the robots.txt file, which controls search engine crawlers' access to specific pages or directories on a website. This tool offers a user-friendly interface to define crawling rules, including allow and disallow directives, to ensure optimal search engine indexing and prevent unwanted content from being indexed. Create and customize your robots.txt file efficiently with this generator to enhance your website's search engine optimization (SEO) efforts."
     },
-    'chart-generator': {
+    'chart-generator-ex': {
         name: 'Chart generator (Beta)',
         icon: 'charts',
-        link: '/chart-generator',
+        link: '/chart-generator-ex',
         description: "Our chart generator tool empowers you to transform raw data into compelling visual insights. Customize, analyze, and share dynamic charts effortlessly, enhancing your data-driven decision-making process."
     },
     'random-data-generator': {
